Rodrigo Grazini - J-B-18-Rodri-Grazi - Changes connection behaviour, adds log4j support and adds resource file for reading pre-defined credentials(file resources/credentials.properties)
# If applied, this commit will...
Add a connect method to a Connection which takes the credentials for the db user and 'creates the session'. Add logging support using log4j
Add properties files for reading user and password defined outside the code with testing purposes

# Why is this change needed?
Prior to this change, the Connections constructors had the responsiblty to make the connection, constructor should never have complicated logic inside them.
Logging was made using standard output inside the code

# How does it adress the issue?
The connection now is handled with a method, constructor only sets up the database URL
Logging is delegated to log4j